+++ codegen/src/lib.rs
@@ -55,6 +55,17 @@ fn create_codegen_file<P, Q>(_: P, codegen_fname: Q) -> Result<(), Box<Error>>
     Ok(())
 }
 
+#[cfg(not(any(feature = "bundle_files", feature = "serve_files")))]
+fn create_codegen_file<P, Q>(_: P, codegen_fname: Q) -> Result<(), Box<Error>>
+    where P: AsRef<Path>,
+          Q: AsRef<Path>
+{
+    // Intentionally trigger a compile time error to force a feature
+    // flag to be used.
+    error = "You are attempting to compile without a required feature flag \
+    being used. You must use one of either `bundle_files` or `serve_files`";
+}
+
 /// Update the codegen file (`codegen_fname`) to include
 /// the `Config`.
 fn include_config<P, Q>(files_dir: P, codegen_fname: Q) -> Result<(), Box<Error>>