+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All user visible changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
+## unreleased
+
+### Added
+
+* Make compile-time error more readable when compiling codegen
+  crate without required feature flag.
+
 ## [0.6.0] - 2018-04-19
 
 ### Changed